requirements.txt
eventlet==0.23.0
psutil==5.4.6
python-qbittorrent==0.3.1
